third_party/blink/web_tests/external/wpt/css/css-sizing/contain-intrinsic-size/auto-009-expected.txt

This is a testharness.js-based test.
FAIL Only recording last remembered inline size assert_equals: Using last remembered inline size - clientHeight expected 1 but got 50
FAIL Only recording last remembered block size assert_equals: Using last remembered block size - clientWidth expected 2 but got 100
Harness: the test ran to completion.
